+++ tools/consumer-fixtures/styled-controls/src/main.ts
@@ -2,12 +2,14 @@ import { Component } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 
 import { CheckboxForms } from './checkbox-forms';
+import { ComboboxForms } from './combobox-forms';
 import { StyledPrimitives } from './controls';
 import { DateInputForms } from './date-input-forms';
 import { StyledInputs } from './inputs';
 import { NumberInputForms } from './number-input-forms';
 import { ComboboxProjection } from './projection';
 import { RadioForms } from './radio-forms';
+import { SelectForms } from './select-forms';
 import { SliderForms } from './slider-forms';
 import { SwitchForms } from './switch-forms';
 import { StyledTable } from './table';
@@ -23,6 +25,8 @@ import { ToggleGroupForms } from './toggle-group-forms';
     CheckboxForms,
     SwitchForms,
     RadioForms,
+    SelectForms,
+    ComboboxForms,
     SliderForms,
     DateInputForms,
     NumberInputForms,
@@ -39,6 +43,8 @@ import { ToggleGroupForms } from './toggle-group-forms';
       <app-checkbox-forms />
       <app-switch-forms />
       <app-radio-forms />
+      <app-select-forms />
+      <app-combobox-forms />
       <app-slider-forms />
       <app-date-input-forms />
       <app-number-input-forms />